refactor(tests): consolidate Milestone Module — 10 files → 4
Closes #3753
