+++ .changeset/3753-consolidate-milestone-tests.md
@@ -1,0 +1,6 @@
+---
+type: patch
+---
+
+<!-- docs-exempt: internal test refactor only — no user-facing surface changed -->
+**Milestone Module test consolidation (#3753)** — 10 test files → 4. All 99 CJS assertions and 11 SDK assertions preserved. `CONTEXT.md` Glossary updated with Milestone Module entry.
